# Improve performance of `Bundler::SpecSet#for` by using hash lookup of handled deps