@deivid-rodriguez 2022-05-14
Oh my god, how did I not realize about this before? 😱 THANK YOU!

If I understand correctly, this improves things more than #5536, and would make #5536 no longer have an effect in performance, similarly to how #5533 made #5532 and #5534 unnecessary?