@deivid-rodriguez 2022-05-16
This looks great to me, I run it against the `rails/rails` repository Gemfile and got the following:

```
➜  rails git:(main) ✗ hyperfine 'BUNDLER_VERSION=2.4.0.hashlookup ruby -rbundler/setup -e1' 'BUNDLER_VERSION=2.4.0.dev ruby -rbundler/setup -e1'
Benchmark 1: BUNDLER_VERSION=2.4.0.hashlookup ruby -rbundler/setup -e1
  Time (mean ± σ):     195.0 ms ±   0.9 ms    [User: 140.2 ms, System: 45.9 ms]
  Range (min … max):   193.1 ms … 196.6 ms    15 runs
 
Benchmark 2: BUNDLER_VERSION=2.4.0.dev ruby -rbundler/setup -e1
  Time (mean ± σ):     203.3 ms ±   0.7 ms    [User: 148.6 ms, System: 45.9 ms]
  Range (min … max):   201.9 ms … 204.2 ms    14 runs
 
Summary
  'BUNDLER_VERSION=2.4.0.hashlookup ruby -rbundler/setup -e1' ran
    1.04 ± 0.01 times faster than 'BUNDLER_VERSION=2.4.0.dev ruby -rbundler/setup -e1'
```

So an extra 4% speed up over current master branch!