@technicalpickles 2022-05-17
> If I understand correctly, this improves things more than https://github.com/rubygems/rubygems/pull/5536, and would make https://github.com/rubygems/rubygems/pull/5536 no longer have an effect in performance, similarly to how https://github.com/rubygems/rubygems/pull/5533 made https://github.com/rubygems/rubygems/pull/5532 and https://github.com/rubygems/rubygems/pull/5534 unnecessary?

I had a chance to test this against https://github.com/technicalpickles/big-gemfile

```
# master
❯ hyperfine "ruby run.rb 2.4.0.dev" --warmup 2                                                                                                      ─╯
Benchmark 1: ruby run.rb 2.4.0.dev
  Time (mean ± σ):     515.6 ms ±  23.8 ms    [User: 304.6 ms, System: 151.9 ms]
  Range (min … max):   482.9 ms … 556.4 ms    10 runs

# this branch
❯ hyperfine "ruby run.rb 2.4.0.dev" --warmup 2                                                                                                      ─╯
Benchmark 1: ruby run.rb 2.4.0.dev
  Time (mean ± σ):     480.8 ms ±  16.5 ms    [User: 278.7 ms, System: 151.4 ms]
  Range (min … max):   457.9 ms … 509.2 ms    10 runs

# this branch + https://github.com/rubygems/rubygems/pull/5536
❯ hyperfine "ruby run.rb 2.4.0.dev" --warmup 2                                                                                                      ─╯
Benchmark 1: ruby run.rb 2.4.0.dev
  Time (mean ± σ):     454.9 ms ±  17.7 ms    [User: 266.7 ms, System: 137.3 ms]
  Range (min … max):   434.4 ms … 480.8 ms    10 runs
```

I cracked open a flamegraph using both the branches, and I'm seeing `LazySpecification#hash` crop up again 😅  cc https://github.com/rubygems/rubygems/pull/5534  ... if we try all 3 branches together...

```
❯ hyperfine "ruby run.rb 2.4.0.dev" --warmup 2                                                                                                      ─╯
Benchmark 1: ruby run.rb 2.4.0.dev
  Time (mean ± σ):     425.1 ms ±  11.6 ms    [User: 233.4 ms, System: 136.2 ms]
  Range (min … max):   412.2 ms … 450.7 ms    10 runs
 ```