@deivid-rodriguez 2022-05-18
I'm not seeing the same results with your sample repository 😮. I get the results that I would expect, namely, this PR improves performance by a lot, but once applied, #5534 and #5536 no longer have any effect.

```
➜  big-gemfile git:(main) ✗ hyperfine 'BUNDLER_VERSION=2.4.0.hash ruby -rbundler/setup -e1' 'BUNDLER_VERSION=2.4.0.hashmemo ruby -rbundler/setup -e1' 'BUNDLER_VERSION=2.4.0.hashtwomemos ruby -rbundler/setup -e1' 'BUNDLER_VERSION=2.4.0.dev ruby -rbundler/setup -e1'
Benchmark 1: BUNDLER_VERSION=2.4.0.hash ruby -rbundler/setup -e1
  Time (mean ± σ):     191.4 ms ±   1.0 ms    [User: 148.0 ms, System: 36.4 ms]
  Range (min … max):   189.8 ms … 193.3 ms    15 runs
 
Benchmark 2: BUNDLER_VERSION=2.4.0.hashmemo ruby -rbundler/setup -e1
  Time (mean ± σ):     190.8 ms ±   1.0 ms    [User: 147.6 ms, System: 36.3 ms]
  Range (min … max):   189.2 ms … 192.4 ms    15 runs
 
Benchmark 3: BUNDLER_VERSION=2.4.0.hashtwomemos ruby -rbundler/setup -e1
  Time (mean ± σ):     190.7 ms ±   0.8 ms    [User: 147.2 ms, System: 36.3 ms]
  Range (min … max):   189.4 ms … 192.3 ms    15 runs
 
Benchmark 4: BUNDLER_VERSION=2.4.0.dev ruby -rbundler/setup -e1
  Time (mean ± σ):     227.1 ms ±   2.6 ms    [User: 183.7 ms, System: 36.3 ms]
  Range (min … max):   225.1 ms … 234.7 ms    12 runs
 
Summary
  'BUNDLER_VERSION=2.4.0.hashtwomemos ruby -rbundler/setup -e1' ran
    1.00 ± 0.01 times faster than 'BUNDLER_VERSION=2.4.0.hashmemo ruby -rbundler/setup -e1'
    1.00 ± 0.01 times faster than 'BUNDLER_VERSION=2.4.0.hash ruby -rbundler/setup -e1'
    1.19 ± 0.01 times faster than 'BUNDLER_VERSION=2.4.0.dev ruby -rbundler/setup -e1'